requirements.txt
openai>=1.30.0
anthropic>=0.25.0
chromadb>=0.5.0
fastapi>=0.111.0
uvicorn>=0.30.0
streamlit>=1.35.0
pdfplumber>=0.11.0
python-dotenv>=1.0.0
pydantic>=2.7.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
cohere>=5.5.0
requests>=2.32.0
aiohttp>=3.9.0
matplotlib>=3.9.0
pandas>=2.2.0
numpy>=1.26.0
